Refactored Pgp.key and merged with mater
