+++ extension/chrome/settings/modules/backup.ts
@@ -11,7 +11,7 @@ import { BrowserMsg } from '../../../js/common/browser/browser-msg.js';
 import { Rules } from '../../../js/common/rules.js';
 import { Lang } from '../../../js/common/lang.js';
 import { Settings } from '../../../js/common/settings.js';
-import { Pgp, KeyInfo } from '../../../js/common/core/pgp.js';
+import { KeyInfo } from '../../../js/common/core/pgp.js';
 import { GoogleAuth } from '../../../js/common/api/google-auth.js';
 import { Buf } from '../../../js/common/core/buf.js';
 import { GMAIL_RECOVERY_EMAIL_SUBJECTS } from '../../../js/common/core/const.js';
@@ -23,6 +23,7 @@ import { KeyImportUi } from './../../../js/common/ui/key_import_ui.js';
 import { Gmail } from '../../../js/common/api/email_provider/gmail/gmail.js';
 import { Ui } from '../../../js/common/browser/ui.js';
 import { ApiErr } from '../../../js/common/api/error/api-error.js';
+import { PgpKey } from '../../../js/common/core/pgp-key.js';
 
 declare const openpgp: typeof OpenPGP;
 
@@ -171,7 +172,7 @@ View.run(class BackupView extends View {
       const [primaryKi] = await Store.keysGet(this.acctEmail, ['primary']);
       Assert.abortAndRenderErrorIfKeyinfoEmpty(primaryKi);
       const { keys: [prv] } = await openpgp.key.readArmored(primaryKi.private);
-      await Pgp.key.encrypt(prv, newPassphrase);
+      await PgpKey.encrypt(prv, newPassphrase);
       await Store.passphraseSave('local', this.acctEmail, primaryKi.longid, newPassphrase);
       await Store.keysAdd(this.acctEmail, prv.armor());
       try {
@@ -306,7 +307,7 @@ View.run(class BackupView extends View {
 
   private isPrivateKeyEncrypted = async (ki: KeyInfo) => {
     const { keys: [prv] } = await openpgp.key.readArmored(ki.private);
-    if (await Pgp.key.decrypt(prv, '', undefined, 'OK-IF-ALREADY-DECRYPTED') === true) {
+    if (await PgpKey.decrypt(prv, '', undefined, 'OK-IF-ALREADY-DECRYPTED') === true) {
       return false;
     }
     return prv.isFullyEncrypted();
@@ -389,7 +390,7 @@ View.run(class BackupView extends View {
   }
 
   private isPassPhraseStrongEnough = async (ki: KeyInfo, passphrase: string) => {
-    const prv = await Pgp.key.read(ki.private);
+    const prv = await PgpKey.read(ki.private);
     if (!prv.isFullyEncrypted()) {
       return false;
     }
@@ -398,7 +399,7 @@ View.run(class BackupView extends View {
       if (!pp) {
         return false;
       }
-      if (await Pgp.key.decrypt(prv, pp) !== true) {
+      if (await PgpKey.decrypt(prv, pp) !== true) {
         await Ui.modal.warning('Pass phrase did not match, please try again.');
         return false;
       }
@@ -413,7 +414,7 @@ View.run(class BackupView extends View {
 
   private setupCreateSimpleAutomaticInboxBackup = async () => {
     const [primaryKi] = await Store.keysGet(this.acctEmail, ['primary']);
-    if (!(await Pgp.key.read(primaryKi.private)).isFullyEncrypted()) {
+    if (!(await PgpKey.read(primaryKi.private)).isFullyEncrypted()) {
       await Ui.modal.warning('Key not protected with a pass phrase, skipping');
       throw new UnreportableError('Key not protected with a pass phrase, skipping');
     }
