+++ extension/chrome/settings/modules/test_passphrase.ts
@@ -6,13 +6,13 @@ import { Store } from '../../../js/common/platform/store.js';
 import { Ui } from '../../../js/common/browser/ui.js';
 import { BrowserMsg } from '../../../js/common/browser/browser-msg.js';
 import { Settings } from '../../../js/common/settings.js';
-import { Pgp } from '../../../js/common/core/pgp.js';
 import { Lang } from '../../../js/common/lang.js';
 import { Assert } from '../../../js/common/assert.js';
 import { initPassphraseToggle } from '../../../js/common/ui/passphrase_ui.js';
 import { Xss } from '../../../js/common/platform/xss.js';
 import { Url } from '../../../js/common/core/common.js';
 import { View } from '../../../js/common/view.js';
+import { PgpKey } from '../../../js/common/core/pgp-key.js';
 
 View.run(class TestPassphrase extends View {
   private readonly acctEmail: string;
@@ -30,7 +30,7 @@ View.run(class TestPassphrase extends View {
     const [keyInfo] = await Store.keysGet(this.acctEmail, ['primary']);
     Assert.abortAndRenderErrorIfKeyinfoEmpty(keyInfo);
     await initPassphraseToggle(['password']);
-    this.primaryKey = await Pgp.key.read(keyInfo.private);
+    this.primaryKey = await PgpKey.read(keyInfo.private);
     if (!this.primaryKey.isFullyEncrypted()) {
       const setUpPpUrl = Url.create('change_passphrase.htm', { acctEmail: this.acctEmail, parentTabId: this.parentTabId });
       Xss.sanitizeRender('#content', `<div class="line">No pass phrase set up yet: <a href="${setUpPpUrl}">set up pass phrase</a></div>`);
@@ -45,7 +45,7 @@ View.run(class TestPassphrase extends View {
   }
 
   private verifyHandler = async () => {
-    if (await Pgp.key.decrypt(this.primaryKey!, String($('#password').val())) === true) {
+    if (await PgpKey.decrypt(this.primaryKey!, String($('#password').val())) === true) {
       Xss.sanitizeRender('#content', `
         <div class="line">${Lang.setup.ppMatchAllSet}</div>
         <div class="line"><button class="button green close" data-test="action-test-passphrase-successful-close">close</button></div>
