+++ extension/js/common/core/mime.ts
@@ -3,13 +3,14 @@
 'use strict';
 
 import { Str, Dict } from './common.js';
-import { Pgp, KeyDetails, DecryptError, VerifyRes } from './pgp.js';
+import { Pgp, DecryptError, VerifyRes } from './pgp.js';
 import { Att, AttMeta } from './att.js';
 import { Catch } from '../platform/catch.js';
 import { requireMimeParser, requireMimeBuilder, requireIso88592 } from '../platform/require.js';
 import { Buf } from './buf.js';
 import { MimeParserNode } from './types/emailjs';
 import { PgpArmor } from './pgp-armor.js';
+import { KeyDetails } from './pgp-key.js';
 
 const MimeParser = requireMimeParser();  // tslint:disable-line:variable-name
 const MimeBuilder = requireMimeBuilder();  // tslint:disable-line:variable-name
